+++ Menu-o-Matic/src/main/groovy/LaunchDialog.groovy
@@ -1,16 +1,18 @@
 package edofro.menuomatic
 
 import groovy.swing.SwingBuilder
-import javax.swing.*
-import java.awt.*
-import java.awt.event.*
+import javax.swing.SwingConstants
+import java.awt.GridLayout
+import java.awt.Insets
+import java.awt.Dimension
+import java.awt.event.WindowFocusListener
+
 
 import groovy.time.TimeCategory
 import groovy.time.TimeDuration
 
 
 import edofro.menuomatic.PackMenu                   as PM
-import edofro.menuomatic.WSE_redux                  as WSE
 import edofro.menuomatic.DialogKeyboardNavigation   as DKBN
     
 import org.freeplane.core.ui.components.UITools     as ui
@@ -35,16 +37,17 @@ class LaunchDialog{
 
 
 
-    // -------------------------- Methods -----------------------------------------------------
+    // region managing dialogs
+
     def static getDialogs(){
         return ui.frame.ownedWindows.findAll{it.name.startsWith(dialogStr) && it.type.toString()=='NORMAL'}
     }
     
     def static showLastUsedDialog(){
         def dialogos = getDialogs()
-        dialogos.max{it.lastUsed}.show()
+        dialogos.max{it.lastUsed}.setVisible(true)
     }
-    
+
     def static listDialogs(){
         def dialogos = getDialogs()
         def titulos = dialogos.sort{it.lastUsed}*.title.reverse()
@@ -53,81 +56,155 @@ class LaunchDialog{
         def titulo = PM.respuestaDialogo(titulos,msg,title)
         dialogos.find{it.title == titulo}?.show()
     }
-    
-    def static showDialogFromMD(MD){
-        md = MD
-        def dialogName = dialogStr + md.title.replace(' ' ,'_')
-        prefDimension = new Dimension(( md.showLabels?100:0) + (md.showIcons?30:0 ) ,md.showIcons?30:25)
-        minDimension = new Dimension(30 ,md.showIcons?30:25)
-        
-        def dialogo = ui.frame.ownedWindows.find{it.name == dialogName && it.type.toString()=='NORMAL'}
-        if( !dialogo ) {
-            c.statusInfo = '------------ new menu dialog created --------------'
-            dialogo = swingBuilder.dialog(
-                title              : md.title,
-                name               : dialogName,
-                modal              : false,
-                locationRelativeTo : ui.frame,
-                minimumSize        : new Dimension(30,70),
-                owner              : ui.frame
-                //defaultCloseOperation: JFrame.DISPOSE_ON_CLOSE,
-            ) {}
-        } else {
-            c.statusInfo = '------------- existing menu dialog reused -------------------'
-            dialogo.getContentPane().removeAll()
-        }        
-        def cols = md.showLabels?1:buttonCols(md.actions.size(),maxButtonsHoriz)
-        dialogo.add(creaContenido(cols)) //TODO: 
-        dialogo.pack()
-        dialogo.show()
-        setLastUsed(dialogo)
-        setWindowFocusListeners(dialogo)
 
-        DKBN.addArrowMoves(dialogo)
-        DKBN.addEscapeAction(dialogo)
-    }
-    
+    // endregion
+
+    // region dialog methods
+
     def static setLastUsed(d){
         if(d.hasProperty('lastUsed')){
             d.lastUsed = new Date()
         } else {
             d.metaClass.lastUsed = new Date()
         }
-
-    
     }
-    
+
     def static setWindowFocusListeners(d){
         d.windowFocusListeners.each{
             d.removeWindowFocusListener(it)
         }
         d.addWindowFocusListener(
-            [windowGainedFocus : { e -> setLastUsed(d)} , 
-             windowLostFocus   : { e -> }] as WindowFocusListener )
-    }
-    
-    def static show(nodo){
-        if(isCustomMenuPack(nodo)){
-            showDialogFromMD(new PM.MenuData(nodo))
-        } else {
-            c.statusInfo = 'selected node is not a customMenu node'
-        }
-        
-    }
-    
-    def static buttonCols(x,cMax){
-        int L = (cMax+x-1)/cMax
-        int c = (x+L-1)/L        
+                [windowGainedFocus : { e -> setLastUsed(d)} ,
+                 windowLostFocus   : { e -> }] as WindowFocusListener )
     }
 
-    def static isCustomMenuPack(n){
+    // endregion
+
+    // region node information
+
+    def static isCustomMenuPack(org.freeplane.api.Node n){
         n.attributes.containsKey(tb.actions)
     }
 
-    def static creaContenido(cols){
+    def static isAutoLaunchMarked(org.freeplane.api.Node n){
+        n.icons.contains('launch')
+    }
+
+    def static isAutoLaunchMenuPack(org.freeplane.api.Node n){
+        isCustomMenuPack(n) && isAutoLaunchMarked(n)
+    }
+
+    // endregion
+    def static launchAutoLaunchCustomMenusFromMap(org.freeplane.features.map.MapModel mapModel){
+        def mapa = ScriptUtils.c().mapLoader(mapModel.getURL()).mindMap
+        if (mapa) launchAutoLaunchCustomMenusFromMap(mapa)
+    }
+	
+    def static launchAutoLaunchCustomMenusFromMap(org.freeplane.api.MindMap mapa){
+        println "|- mindMap: '${mapa.name}'"
+        def nodos = mapa.root.find{isAutoLaunchMenuPack(it)}
+        println "|  - ${nodos*.text}"
+        nodos.each{
+            show(it)
+        }
+    }	
+	
+	
+    def static show(nodo){
+        if(isCustomMenuPack(nodo)){
+            md = new PM.MenuData(nodo)
+            prefDimension = new Dimension(( md.showLabels?100:0) + (md.showIcons?30:0 ) ,md.showIcons?30:25)
+            minDimension = new Dimension(30 ,md.showIcons?30:25)
+            def panelName = md.title.replace(' ' ,'_')
+            showDialog(panelName)
+        } else {
+            c.statusInfo = 'selected node is not a customMenu node'
+        }
+    }
+
+    // region creating Tab menu
+
+    def static showTabMenu(tabName, panelName){
+        return null
+    }
+
+    // endregion
+
+    //region creating/showing dialog menu
+
+    def static showDialogFromMD(MD){
+        md = MD
+        prefDimension = new Dimension(( md.showLabels?100:0) + (md.showIcons?30:0 ) ,md.showIcons?30:25)
+        minDimension = new Dimension(30 ,md.showIcons?30:25)
+        def panelName = md.title.replace(' ' ,'_')
+        showDialog(panelName)
+    }
+
+    def static showDialog(panelName){
+        def dialogo = getEmptyDialog(dialogStr + panelName)
+        dialogo.add(panelContenido(panelName))
+        dialogo.pack()
+        dialogo.setVisible(true)
+        setLastUsed(dialogo)
+        setWindowFocusListeners(dialogo)
+
+        DKBN.addArrowMoves(dialogo)
+        DKBN.addEscapeAction(dialogo)
+    }
+
+    def static getEmptyDialog(String dialogName){
+        def dialogo = ui.frame.ownedWindows.find{it.name == dialogName && it.type.toString()=='NORMAL'}
+        if( !dialogo ) {
+            c.statusInfo = '------------ new menu dialog created --------------'
+            dialogo = swingBuilder.dialog(
+                    title              : md.title,
+                    name               : dialogName,
+                    modal              : false,
+                    locationRelativeTo : ui.frame,
+                    minimumSize        : new Dimension(30,70),
+                    owner              : ui.frame
+                    //defaultCloseOperation: JFrame.DISPOSE_ON_CLOSE,
+            ) {}
+        } else {
+            c.statusInfo = '------------- existing menu dialog reused -------------------'
+            dialogo.getContentPane().removeAll()
+        }
+        return dialogo
+    }
+
+    // endregion
+
+    //region creating panel
+
+    def static getCols(){
+        md.showLabels?1:buttonCols(md.actions.size(),maxButtonsHoriz)
+    }
+
+    def static buttonCols(x,cMax){
+        int L = (cMax+x-1)/cMax
+        int c = (x+L-1)/L
+    }
+
+    def static panelContenido(){
+        def cols = getCols()
+        def panelName = 'miPanel'
+        return panelContenido(cols, panelName)
+    }
+
+    def static panelContenido(Integer cols){
+        def panelName = 'miPanel'
+        return panelContenido(cols, panelName)
+    }
+
+    def static panelContenido(String panelName){
+        def cols = getCols()
+        return panelContenido(cols, panelName)
+    }
+
+    def static panelContenido(Integer cols, String panelName){
         return swingBuilder.panel(
-                layout : new GridLayout(0,cols),
-                name   : 'miPanel'
+                layout : new GridLayout(0, cols),
+                name   : panelName
             ){
     //           toolBar(
     //           layout: new FlowLayout(FlowLayout.LEFT, 0, 0)
@@ -142,6 +219,9 @@ class LaunchDialog{
         }
     }
 
+    // endregion
+
+    // region creating buttons
     def static creaBoton(acc, i){
         if(acc.startsWith(scriptStr)){
             return creaBotonDesdeScript(acc,i)
@@ -171,19 +251,16 @@ class LaunchDialog{
     }
 
     def static creaBotonDesdeUI(acc, i){
-        def text        = md.showLabels?textoLabel(md.labels[i]):null
-        def icon        = md.showIcons?menuUtils.getMenuItemIcon(md.icons[i]):null
+        def text = md.showLabels?textoLabel(md.labels[i]):null
+        def icon = md.showIcons?menuUtils.getMenuItemIcon(md.icons[i]):null
         def toolTipText = md.labels[i]
-        def minD        = minDimension
+        def minD = minDimension
         def actionPerformed = {
                 menuUtils.executeMenuItems([acc])
                 if (md.focusMap) DKBN.focusMap()
             }
         return nuevoBoton(text,icon,toolTipText,minD, actionPerformed)
     }
-
-
-//----------------------------------------------------
 
     def static textoLabel(texto) {
       return textUtils.getShortText(texto,md.maxTextLength,'.')
@@ -203,5 +280,5 @@ class LaunchDialog{
         )
         return boton
     }
-    
+    // endregion
 }